requirements.txt
python-telegram-bot
telethon
httpx==0.27.2
requests
tqdm
m3u8
pycryptodome
python-dotenv
